refactor: split app and improve build config
