+++ README.md
@@ -1,2 +1,20 @@
-# kpi-retention
+# KPI Retention
 
+Simple Vite + React app to track driver KPIs and retention.
+
+## Quickstart
+```bash
+npm install
+npm run dev
+```
+
+## Build / Preview
+```bash
+npm run build
+npm run preview
+```
+
+## Tech Stack
+- Vite
+- React
+- Tailwind CSS
