+++ tailwind.config.js
@@ -1,9 +1,11 @@
 /** @type {import('tailwindcss').Config} */
+import forms from '@tailwindcss/forms';
+
 export default {
   content: [
-    "./index.html",
-    "./src/**/*.{js,ts,jsx,tsx}",
+    './index.html',
+    './src/**/*.{js,ts,jsx,tsx}',
   ],
   theme: { extend: {} },
-  plugins: [],
-}
+  plugins: [forms],
+};
